Add method to close incidents and refactor Slack adapter package structure
